chore: Refactor command restqa run
