+++ src/cli/run.js
@@ -2,7 +2,7 @@ const path = require("path");
 const fs = require("fs");
 const cucumber = require("@cucumber/cucumber");
 
-const {program: cliProgram} = require("../../bin/program");
+const Initialize= require("../core/initialize");
 const logger = require("../utils/logger");
 
 module.exports = async function (opt, program = {}) {
@@ -12,7 +12,6 @@ module.exports = async function (opt, program = {}) {
     tags = [],
     stream = process.stdout,
     args,
-    exec: command,
     skipInit = false
   } = opt;
 
@@ -54,7 +53,7 @@ module.exports = async function (opt, program = {}) {
   config = config || path.join(process.cwd(), ".restqa.yml");
   if (!fs.existsSync(config)) {
     if (!skipInit) {
-      await cliProgram.parseAsync([process.argv[0], process.argv[1], "init"]);
+      await Initialize();
     } else {
       return Promise.reject(
         new TypeError(`The configuration file "${config}" doesn't exist.`)
@@ -64,8 +63,7 @@ module.exports = async function (opt, program = {}) {
 
   global.restqaOptions = {
     config,
-    env,
-    command
+    env
   };
 
   // TODO : Add extra cucumber parameters from config file
